renaming facts to info (#54432)
(cherry picked from commit c5edffcdb8042f041eee41e2f11ccd2e4567e3ee)
