+++ lib/ansible/modules/cloud/azure/azure_rm_mariadbdatabase_info.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 #
-# Copyright (c) 2019 Zim Kalinowski, (@zikalino)
+# Copyright (c) 2017 Zim Kalinowski, <zikalino@microsoft.com>
 # Copyright (c) 2019 Matti Ranta, (@techknowlogick)
 #
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
@@ -16,11 +16,11 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 
 DOCUMENTATION = '''
 ---
-module: azure_rm_mariadbconfiguration_facts
+module: azure_rm_mariadbdatabase_info
 version_added: "2.8"
-short_description: Get Azure MariaDB Configuration facts.
+short_description: Get Azure MariaDB Database facts.
 description:
-    - Get facts of Azure MariaDB Configuration.
+    - Get facts of MariaDB Database.
 
 options:
     resource_group:
@@ -33,7 +33,7 @@ options:
         required: True
     name:
         description:
-            - Setting name.
+            - The name of the database.
 
 extends_documentation_fragment:
     - azure
@@ -45,62 +45,67 @@ author:
 '''
 
 EXAMPLES = '''
-  - name: Get specific setting of MariaDB Server
-    azure_rm_mariadbconfiguration_facts:
+  - name: Get instance of MariaDB Database
+    azure_rm_mariadbdatabase_info:
       resource_group: myResourceGroup
-      server_name: testserver
-      name: deadlock_timeout
+      server_name: server_name
+      name: database_name
 
-  - name: Get all settings of MariaDB Server
-    azure_rm_mariadbconfiguration_facts:
+  - name: List instances of MariaDB Database
+    azure_rm_mariadbdatabase_info:
       resource_group: myResourceGroup
       server_name: server_name
 '''
 
 RETURN = '''
-settings:
-    description: A list of dictionaries containing MariaDB Server settings.
+databases:
+    description: A list of dictionaries containing facts for MariaDB Databases.
     returned: always
     type: complex
     contains:
         id:
             description:
-                - Setting resource ID
+                - Resource ID
             returned: always
             type: str
-            sample: "/subscriptions/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/resourceGroups/myResourceGroup/providers/Microsoft.DBforMariaDB/servers/testserver
-                     /configurations/deadlock_timeout"
+            sample: "/subscriptions/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/resourceGroups/myResourceGroup/providers/Microsoft.DBforMariaDB/servers/testser
+                    ver/databases/db1"
+        resource_group:
+            description:
+                - Resource group name.
+            returned: always
+            type: str
+            sample: testrg
+        server_name:
+            description:
+                - Server name.
+            returned: always
+            type: str
+            sample: testserver
         name:
             description:
-                - Setting name.
+                - Resource name.
             returned: always
             type: str
-            sample: deadlock_timeout
-        value:
+            sample: db1
+        charset:
             description:
-                - Setting value.
-            returned: always
-            type: raw
-            sample: 1000
-        description:
-            description:
-                - Description of the configuration.
+                - The charset of the database.
             returned: always
             type: str
-            sample: Deadlock timeout.
-        source:
+            sample: UTF8
+        collation:
             description:
-                - Source of the configuration.
+                - The collation of the database.
             returned: always
             type: str
-            sample: system-default
+            sample: English_United States.1252
 '''
 
 from ansible.module_utils.azure_rm_common import AzureRMModuleBase
 
 try:
     from msrestazure.azure_exceptions import CloudError
-    from msrestazure.azure_operation import AzureOperationPoller
     from azure.mgmt.rdbms.mariadb import MariaDBManagementClient
     from msrest.serialization import Model
 except ImportError:
@@ -108,7 +113,7 @@ except ImportError:
     pass
 
 
-class AzureRMMariaDbConfigurationFacts(AzureRMModuleBase):
+class AzureRMMariaDbDatabaseFacts(AzureRMModuleBase):
     def __init__(self):
         # define user inputs into argument
         self.module_arg_spec = dict(
@@ -125,40 +130,37 @@ class AzureRMMariaDbConfigurationFacts(AzureRMModuleBase):
             )
         )
         # store the results of the module operation
-        self.results = dict(changed=False)
-        self.mgmt_client = None
+        self.results = dict(
+            changed=False
+        )
         self.resource_group = None
         self.server_name = None
         self.name = None
-        super(AzureRMMariaDbConfigurationFacts, self).__init__(self.module_arg_spec, supports_tags=False)
+        super(AzureRMMariaDbDatabaseFacts, self).__init__(self.module_arg_spec, supports_tags=False)
 
     def exec_module(self, **kwargs):
         for key in self.module_arg_spec:
             setattr(self, key, kwargs[key])
-        self.mgmt_client = self.get_mgmt_svc_client(MariaDBManagementClient,
-                                                    base_url=self._cloud_environment.endpoints.resource_manager)
 
-        if self.name is not None:
-            self.results['settings'] = self.get()
-        else:
-            self.results['settings'] = self.list_by_server()
+        if (self.resource_group is not None and
+                self.server_name is not None and
+                self.name is not None):
+            self.results['databases'] = self.get()
+        elif (self.resource_group is not None and
+              self.server_name is not None):
+            self.results['databases'] = self.list_by_server()
         return self.results
 
     def get(self):
-        '''
-        Gets facts of the specified MariaDB Configuration.
-
-        :return: deserialized MariaDB Configurationinstance state dictionary
-        '''
         response = None
         results = []
         try:
-            response = self.mgmt_client.configurations.get(resource_group_name=self.resource_group,
-                                                           server_name=self.server_name,
-                                                           configuration_name=self.name)
+            response = self.mariadb_client.databases.get(resource_group_name=self.resource_group,
+                                                         server_name=self.server_name,
+                                                         database_name=self.name)
             self.log("Response : {0}".format(response))
         except CloudError as e:
-            self.log('Could not get facts for Configurations.')
+            self.log('Could not get facts for Databases.')
 
         if response is not None:
             results.append(self.format_item(response))
@@ -166,19 +168,14 @@ class AzureRMMariaDbConfigurationFacts(AzureRMModuleBase):
         return results
 
     def list_by_server(self):
-        '''
-        Gets facts of the specified MariaDB Configuration.
-
-        :return: deserialized MariaDB Configurationinstance state dictionary
-        '''
         response = None
         results = []
         try:
-            response = self.mgmt_client.configurations.list_by_server(resource_group_name=self.resource_group,
-                                                                      server_name=self.server_name)
+            response = self.mariadb_client.databases.list_by_server(resource_group_name=self.resource_group,
+                                                                    server_name=self.server_name)
             self.log("Response : {0}".format(response))
         except CloudError as e:
-            self.log('Could not get facts for Configurations.')
+            self.fail("Error listing for server {0} - {1}".format(self.server_name, str(e)))
 
         if response is not None:
             for item in response:
@@ -191,17 +188,15 @@ class AzureRMMariaDbConfigurationFacts(AzureRMModuleBase):
         d = {
             'resource_group': self.resource_group,
             'server_name': self.server_name,
-            'id': d['id'],
             'name': d['name'],
-            'value': d['value'],
-            'description': d['description'],
-            'source': d['source']
+            'charset': d['charset'],
+            'collation': d['collation']
         }
         return d
 
 
 def main():
-    AzureRMMariaDbConfigurationFacts()
+    AzureRMMariaDbDatabaseFacts()
 
 
 if __name__ == '__main__':
