+++ lib/ansible/modules/cloud/azure/azure_rm_mysqlfirewallrule_info.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 #
-# Copyright (c) 2017 Zim Kalinowski, <zikalino@microsoft.com>
+# Copyright (c) 2018 Zim Kalinowski, <zikalino@microsoft.com>
 #
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
@@ -15,16 +15,16 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 
 DOCUMENTATION = '''
 ---
-module: azure_rm_sqlfirewallrule_facts
+module: azure_rm_mysqlfirewallrule_info
 version_added: "2.8"
-short_description: Get Azure SQL Firewall Rule facts.
+short_description: Get Azure MySQL Firewall Rule facts.
 description:
-    - Get facts of SQL Firewall Rule.
+    - Get facts of Azure MySQL Firewall Rule.
 
 options:
     resource_group:
         description:
-            - The name of the resource group that contains the server.
+            - The name of the resource group.
         required: True
     server_name:
         description:
@@ -32,7 +32,7 @@ options:
         required: True
     name:
         description:
-            - The name of the firewall rule.
+            - The name of the server firewall rule.
 
 extends_documentation_fragment:
     - azure
@@ -43,21 +43,21 @@ author:
 '''
 
 EXAMPLES = '''
-  - name: Get instance of SQL Firewall Rule
-    azure_rm_sqlfirewallrule_facts:
+  - name: Get instance of MySQL Firewall Rule
+    azure_rm_mysqlfirewallrule_info:
       resource_group: myResourceGroup
-      server_name: testserver
-      name: testrule
+      server_name: server_name
+      name: firewall_rule_name
 
-  - name: List instances of SQL Firewall Rule
-    azure_rm_sqlfirewallrule_facts:
+  - name: List instances of MySQL Firewall Rule
+    azure_rm_mysqlfirewallrule_info:
       resource_group: myResourceGroup
-      server_name: testserver
+      server_name: server_name
 '''
 
 RETURN = '''
 rules:
-    description: A list of dict results containing the facts for matching SQL firewall rules.
+    description: A list of dictionaries containing facts for MySQL Firewall Rule.
     returned: always
     type: complex
     contains:
@@ -66,53 +66,47 @@ rules:
                 - Resource ID
             returned: always
             type: str
-            sample: "/subscriptions/xxxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/resourceGroups/myResourceGroup/providers/Microsoft.Sql/servers/testser
-                    ver/firewallRules/testrule"
-        resource_group:
-            description:
-                - Resource group name.
-            returned: always
-            type: str
-            sample: testgroup
+            sample: "/subscriptions/xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx/resourceGroups/TestGroup/providers/Microsoft.DBforMySQL/servers/testserver/fire
+                    wallRules/rule1"
         server_name:
             description:
-                - SQL server name.
+                - The name of the server.
             returned: always
             type: str
             sample: testserver
         name:
             description:
-                - Firewall rule name.
+                - Resource name.
             returned: always
             type: str
-            sample: testrule
+            sample: rule1
         start_ip_address:
             description:
-                - The start IP address of the firewall rule.
+                - The start IP address of the MySQL firewall rule.
             returned: always
             type: str
-            sample: 10.0.0.1
+            sample: 10.0.0.16
         end_ip_address:
             description:
-                - The start IP address of the firewall rule.
+                - The end IP address of the MySQL firewall rule.
             returned: always
             type: str
-            sample: 10.0.0.5
+            sample: 10.0.0.18
 '''
 
 from ansible.module_utils.azure_rm_common import AzureRMModuleBase
 
 try:
     from msrestazure.azure_exceptions import CloudError
-    from msrest.polling import LROPoller
-    from azure.mgmt.sql import SqlManagementClient
+    from msrestazure.azure_operation import AzureOperationPoller
+    from azure.mgmt.rdbms.mysql import MySQLManagementClient
     from msrest.serialization import Model
 except ImportError:
     # This is handled in azure_rm_common
     pass
 
 
-class AzureRMSqlFirewallRuleFacts(AzureRMModuleBase):
+class AzureRMMySqlFirewallRuleFacts(AzureRMModuleBase):
     def __init__(self):
         # define user inputs into argument
         self.module_arg_spec = dict(
@@ -132,14 +126,17 @@ class AzureRMSqlFirewallRuleFacts(AzureRMModuleBase):
         self.results = dict(
             changed=False
         )
+        self.mgmt_client = None
         self.resource_group = None
         self.server_name = None
         self.name = None
-        super(AzureRMSqlFirewallRuleFacts, self).__init__(self.module_arg_spec, supports_tags=False)
+        super(AzureRMMySqlFirewallRuleFacts, self).__init__(self.module_arg_spec, supports_tags=False)
 
     def exec_module(self, **kwargs):
         for key in self.module_arg_spec:
             setattr(self, key, kwargs[key])
+        self.mgmt_client = self.get_mgmt_svc_client(MySQLManagementClient,
+                                                    base_url=self._cloud_environment.endpoints.resource_manager)
 
         if (self.name is not None):
             self.results['rules'] = self.get()
@@ -148,17 +145,12 @@ class AzureRMSqlFirewallRuleFacts(AzureRMModuleBase):
         return self.results
 
     def get(self):
-        '''
-        Gets facts of the specified SQL Firewall Rule.
-
-        :return: deserialized SQL Firewall Ruleinstance state dictionary
-        '''
         response = None
         results = []
         try:
-            response = self.sql_client.firewall_rules.get(resource_group_name=self.resource_group,
-                                                          server_name=self.server_name,
-                                                          firewall_rule_name=self.name)
+            response = self.mgmt_client.firewall_rules.get(resource_group_name=self.resource_group,
+                                                           server_name=self.server_name,
+                                                           firewall_rule_name=self.name)
             self.log("Response : {0}".format(response))
         except CloudError as e:
             self.log('Could not get facts for FirewallRules.')
@@ -169,16 +161,11 @@ class AzureRMSqlFirewallRuleFacts(AzureRMModuleBase):
         return results
 
     def list_by_server(self):
-        '''
-        Gets facts of the specified SQL Firewall Rule.
-
-        :return: deserialized SQL Firewall Ruleinstance state dictionary
-        '''
         response = None
         results = []
         try:
-            response = self.sql_client.firewall_rules.list_by_server(resource_group_name=self.resource_group,
-                                                                     server_name=self.server_name)
+            response = self.mgmt_client.firewall_rules.list_by_server(resource_group_name=self.resource_group,
+                                                                      server_name=self.server_name)
             self.log("Response : {0}".format(response))
         except CloudError as e:
             self.log('Could not get facts for FirewallRules.')
@@ -192,8 +179,8 @@ class AzureRMSqlFirewallRuleFacts(AzureRMModuleBase):
     def format_item(self, item):
         d = item.as_dict()
         d = {
-            'id': d['id'],
             'resource_group': self.resource_group,
+            'id': d['id'],
             'server_name': self.server_name,
             'name': d['name'],
             'start_ip_address': d['start_ip_address'],
@@ -203,7 +190,7 @@ class AzureRMSqlFirewallRuleFacts(AzureRMModuleBase):
 
 
 def main():
-    AzureRMSqlFirewallRuleFacts()
+    AzureRMMySqlFirewallRuleFacts()
 
 
 if __name__ == '__main__':
