+++ test/integration/targets/azure_rm_mysqlserver/aliases
@@ -5,3 +5,6 @@ azure_rm_mysqlserver_facts
 azure_rm_mysqldatabase
 azure_rm_mysqldatabase_facts
 azure_rm_mysqlfirewallrule
+azure_rm_mysqlfirewallrule_info
+azure_rm_mysqlconfiguration
+azure_rm_mysqlconfiguration_info
