+++ test/integration/targets/azure_rm_postgresqlserver/aliases
@@ -5,5 +5,6 @@ azure_rm_postgresqlserver_facts
 azure_rm_postgresqldatabase
 azure_rm_postgresqldatabase_facts
 azure_rm_postgresqlfirewallrule
-azure_rm_postgresqlfirewallrule_facts
-azure_rm_postgresqlserverconfiguration_facts
+azure_rm_postgresqlfirewallrule_info
+azure_rm_postgresqlserverconfiguration
+azure_rm_postgresqlserverconfiguration_info
