+++ test/integration/targets/azure_rm_postgresqlserver/tasks/main.yml
@@ -374,7 +374,7 @@
     end_ip_address: 172.28.10.138
 
 - name: Gather facts PostgreSQL Firewall Rule
-  azure_rm_postgresqlfirewallrule_facts:
+  azure_rm_postgresqlfirewallrule_info:
     resource_group: "{{ resource_group }}"
     server_name: postgresqlsrv{{ rpfx }}
     name: firewallrule{{ rpfx }}
@@ -391,7 +391,7 @@
       - "output.rules | length == 1"
 
 - name: Gather facts PostgreSQL Firewall Rule
-  azure_rm_postgresqlfirewallrule_facts:
+  azure_rm_postgresqlfirewallrule_info:
     resource_group: "{{ resource_group }}"
     server_name: postgresqlsrv{{ rpfx }}
   register: output
@@ -455,7 +455,7 @@
     state: absent
 
 - name: Gather facts PostgreSQL Firewall Rule
-  azure_rm_postgresqlfirewallrule_facts:
+  azure_rm_postgresqlfirewallrule_info:
     resource_group: "{{ resource_group }}"
     server_name: postgresqlsrv{{ rpfx }}
     name: firewallrule{{ rpfx }}
@@ -531,7 +531,7 @@
       - not output.changed
 
 - name: Gather facts PostgreSQL Configuration
-  azure_rm_postgresqlconfiguration_facts:
+  azure_rm_postgresqlconfiguration_info:
     resource_group: "{{ resource_group }}"
     server_name: postgresqlsrv{{ rpfx }}
     name: deadlock_timeout
@@ -548,7 +548,7 @@
       - output.settings | length == 1
 
 - name: Gather facts PostgreSQL Configuration
-  azure_rm_postgresqlconfiguration_facts:
+  azure_rm_postgresqlconfiguration_info:
     resource_group: "{{ resource_group }}"
     server_name: postgresqlsrv{{ rpfx }}
   register: output
